requirements.txt
Django==4.2.6
django-cors-headers==3.13.0
django-filter==22.1
django_jazzmin==2.6.0
djangorestframework==3.14.0
drf-yasg==1.21.7
Pillow==9.2.0
pytz==2022.4